Refactor into individual files
